Revert "Sort testnets my mainnet' tvl (#2167)"
This reverts commit 8ff7668ee4c8bf1c29c1b7961bb3f79659e36af4.
